Move mode-matrix and continuum_analysis into package directory
